+++ frontend/src/services/auth.ts
@@ -6,7 +6,7 @@ export async function login({
   data: { email: string; password: string };
 }) {
   try {
-    const response = await axios.post("http://localhost:8080/auth/login", data);
+    const response = await axios.post("http://localhost:3000/auth/login", data);
     localStorage.setItem("token", response.data.token);
     localStorage.setItem("session", JSON.stringify(response.data.user));
     return response;
